XXX: rename fancyfunc (other names welcome?)
